separate welcome docs into a component
